+++ source/common/http/conn_manager_impl.h
@@ -653,6 +653,7 @@ private:
   void onDrainTimeout();
   void startDrainSequence();
   Tracing::HttpTracer& tracer() { return http_context_.tracer(); }
+  void handleCodecException(const char* error, Network::ConnectionCloseType close_type);
 
   enum class DrainState { NotDraining, Draining, Closing };