+++ test/integration/http2_integration_test.cc
@@ -1,5 +1,6 @@
 #include "test/integration/http2_integration_test.h"
 
+#include <algorithm>
 #include <string>
 
 #include "common/buffer/buffer_impl.h"
@@ -1424,6 +1425,184 @@ TEST_P(Http2RingHashIntegrationTest, CookieRoutingWithCookieWithTtlSet) {
             response.headers().get(Http::LowerCaseString("x-served-by"))->value().getStringView()));
       });
   EXPECT_EQ(served_by.size(), 1);
+}
+
+namespace {
+const int64_t TransmitThreshold = 100 * 1024 * 1024;
+} // namespace
+
+void Http2FloodMitigationTest::setNetworkConnectionBufferSize() {
+  // nghttp2 library has its own internal mitigation for outbound control frames. The mitigation is
+  // trigerred when there are more than 10000 PING or SETTINGS frames with ACK flag in the nghttp2
+  // internal outbound queue. It is possible to trigger this mitigation in nghttp2 before triggering
+  // Envoy's own flood mitigation. This can happen when a buffer larger enough to contain over 10K
+  // PING or SETTINGS frames is dispatched to the nghttp2 library. To prevent this from happening
+  // the network connection receive buffer needs to be smaller than 90Kb (which is 10K SETTINGS
+  // frames). Set it to the arbitrarily chosen value of 32K.
+  config_helper_.addConfigModifier([](envoy::config::bootstrap::v2::Bootstrap& bootstrap) -> void {
+    RELEASE_ASSERT(bootstrap.mutable_static_resources()->listeners_size() >= 1, "");
+    auto* listener = bootstrap.mutable_static_resources()->mutable_listeners(0);
+    listener->mutable_per_connection_buffer_limit_bytes()->set_value(32 * 1024);
+  });
+}
+
+void Http2FloodMitigationTest::beginSession() {
+  setDownstreamProtocol(Http::CodecClient::Type::HTTP2);
+  setUpstreamProtocol(FakeHttpConnection::Type::HTTP2);
+  // set lower outbound frame limits to make tests run faster
+  config_helper_.setOutboundFramesLimits(1000, 100);
+  initialize();
+  tcp_client_ = makeTcpConnection(lookupPort("http"));
+  startHttp2Session();
+}
+
+Http2Frame Http2FloodMitigationTest::readFrame() {
+  Http2Frame frame;
+  tcp_client_->waitForData(frame.HeaderSize);
+  frame.setHeader(tcp_client_->data());
+  tcp_client_->clearData(frame.HeaderSize);
+  auto len = frame.payloadSize();
+  if (len) {
+    tcp_client_->waitForData(len);
+    frame.setPayload(tcp_client_->data());
+    tcp_client_->clearData(len);
+  }
+  return frame;
+}
+
+void Http2FloodMitigationTest::sendFame(const Http2Frame& frame) {
+  ASSERT_TRUE(tcp_client_->connected());
+  tcp_client_->write(std::string(frame), false, false);
+}
+
+void Http2FloodMitigationTest::startHttp2Session() {
+  tcp_client_->write(Http2Frame::Preamble, false, false);
+
+  // Send empty initial SETTINGS frame.
+  auto settings = Http2Frame::makeEmptySettingsFrame();
+  tcp_client_->write(std::string(settings), false, false);
+
+  // Read initial SETTINGS frame from the server.
+  readFrame();
+
+  // Send an SETTINGS ACK.
+  settings = Http2Frame::makeEmptySettingsFrame(Http2Frame::SettingsFlags::ACK);
+  tcp_client_->write(std::string(settings), false, false);
+
+  // read pending SETTINGS and WINDOW_UPDATE frames
+  readFrame();
+  readFrame();
+}
+
+// Verify that the server detects the flood of the given frame.
+void Http2FloodMitigationTest::floodServer(const Http2Frame& frame) {
+  config_helper_.setBufferLimits(1024, 1024); // Set buffer limits upstream and downstream.
+  beginSession();
+
+  // pack the as many frames as we can into 16k buffer
+  const int FrameCount = (16 * 1024) / frame.size();
+  std::vector<char> buf(FrameCount * frame.size());
+  for (auto pos = buf.begin(); pos != buf.end();) {
+    pos = std::copy(frame.begin(), frame.end(), pos);
+  }
+
+  tcp_client_->readDisable(true);
+  int64_t total_bytes_sent = 0;
+  // If the flood protection is not working this loop will keep going
+  // forever until it is killed by blaze timer or run out of memory.
+  // Add early stop if we have sent more than 100M of frames, as it this
+  // point it is obvious something is wrong.
+  while (total_bytes_sent < TransmitThreshold && tcp_client_->connected()) {
+    tcp_client_->write({buf.begin(), buf.end()}, false, false);
+    total_bytes_sent += buf.size();
+  }
+
+  EXPECT_LE(total_bytes_sent, TransmitThreshold) << "Flood mitigation is broken.";
+  EXPECT_EQ(1, test_server_->counter("http2.outbound_control_flood")->value());
+  // Verify that connection was closed abortively
+  EXPECT_EQ(0,
+            test_server_->counter("http.config_test.downstream_cx_delayed_close_timeout")->value());
+}
+
+// Verify that the server detects the flood using specified request parameters.
+void Http2FloodMitigationTest::floodServer(absl::string_view host, absl::string_view path,
+                                           Http2Frame::ResponseStatus expected_http_status) {
+  uint32_t request_idx = 0;
+  auto request = Http2Frame::makeRequest(request_idx, host, path);
+  sendFame(request);
+  auto frame = readFrame();
+  EXPECT_EQ(Http2Frame::Type::HEADERS, frame.type());
+  EXPECT_EQ(expected_http_status, frame.responseStatus());
+  tcp_client_->readDisable(true);
+  uint64_t total_bytes_sent = 0;
+  while (total_bytes_sent < TransmitThreshold && tcp_client_->connected()) {
+    request = Http2Frame::makeRequest(++request_idx, host, path);
+    sendFame(request);
+    total_bytes_sent += request.size();
+  }
+  EXPECT_LE(total_bytes_sent, TransmitThreshold) << "Flood mitigation is broken.";
+  EXPECT_EQ(1, test_server_->counter("http2.outbound_flood")->value());
+  // Verify that connection was closed abortively
+  EXPECT_EQ(0,
+            test_server_->counter("http.config_test.downstream_cx_delayed_close_timeout")->value());
+}
+
+INSTANTIATE_TEST_SUITE_P(IpVersions, Http2FloodMitigationTest,
+                         testing::ValuesIn(TestEnvironment::getIpVersionsForTest()),
+                         TestUtility::ipTestParamsToString);
+
+TEST_P(Http2FloodMitigationTest, Ping) {
+  setNetworkConnectionBufferSize();
+  floodServer(Http2Frame::makePingFrame());
+}
+
+TEST_P(Http2FloodMitigationTest, Settings) {
+  setNetworkConnectionBufferSize();
+  floodServer(Http2Frame::makeEmptySettingsFrame());
+}
+
+// Verify that the server can detect flood of internally generated 404 responses.
+TEST_P(Http2FloodMitigationTest, 404) {
+  // Change the default route to be restrictive, and send a request to a non existent route.
+  config_helper_.setDefaultHostAndRoute("foo.com", "/found");
+  beginSession();
+
+  // Send requests to a non existent path to generate 404s
+  floodServer("host", "/notfound", Http2Frame::ResponseStatus::_404);
+}
+
+// Verify that the server can detect flood of DATA frames
+TEST_P(Http2FloodMitigationTest, Data) {
+  // Set large buffer limits so the test is not affected by the flow control.
+  config_helper_.setBufferLimits(1024 * 1024 * 1024, 1024 * 1024 * 1024);
+  autonomous_upstream_ = true;
+  beginSession();
+  fake_upstreams_[0]->set_allow_unexpected_disconnects(true);
+
+  floodServer("host", "/test/long/url", Http2Frame::ResponseStatus::_200);
+}
+
+// Verify that the server can detect flood of RST_STREAM frames.
+TEST_P(Http2FloodMitigationTest, RST_STREAM) {
+  beginSession();
+
+  int i = 0;
+  auto request = Http::Http2::Http2Frame::makeMalformedRequest(i);
+  sendFame(request);
+  auto response = readFrame();
+  // Make sure we've got RST_STREAM from the server
+  EXPECT_EQ(Http2Frame::Type::RST_STREAM, response.type());
+  uint64_t total_bytes_sent = 0;
+  while (total_bytes_sent < TransmitThreshold && tcp_client_->connected()) {
+    request = Http::Http2::Http2Frame::makeMalformedRequest(++i);
+    sendFame(request);
+    total_bytes_sent += request.size();
+  }
+  EXPECT_LE(total_bytes_sent, TransmitThreshold) << "Flood mitigation is broken.";
+  EXPECT_EQ(1, test_server_->counter("http2.outbound_control_flood")->value());
+  // Verify that connection was closed abortively
+  EXPECT_EQ(0,
+            test_server_->counter("http.config_test.downstream_cx_delayed_close_timeout")->value());
 }
 
 } // namespace Envoy